+++ CHANGELOG.md
@@ -15,6 +15,7 @@ If needed, you can also manually update this file (provided the general structur
 - Update color taxonomy.
 - Removed VPN specific brand colors.
 - Updated Splash/Welcome/SignIn/SignUp Screens. Added SplashScreen API support (https://developer.android.com/guide/topics/ui/splash-screen). Client MainActivity migration: Call `installSplashScreen()` before `onCreate`, and optionally `setKeepOnScreenCondition` to minimize transition between Splash and next Activity.
+- Add updated accent colors, clients must use them instead of their current ones.
 
 ## [Unreleased]